requirements.txt
Django==5.2.5
django-ckeditor==6.7.3
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.1